+++ CHANGES.rst
@@ -7,14 +7,21 @@ Major changes
 * New encoder: :class:`DatetimeEncoder` can transform a datetime column into several numerical
     columns (year, month, day, hour, minute, second, ...). It is now the default transformer used
     in the :class:`SuperVectorizer` for datetime columns.
-* The :class:`SuperVectorizer` has seen some major improvements and bug fixes.
+* The :class:`SuperVectorizer` has seen some major improvements and bug fixes
+  - Fixes the automatic casting logic in ``transform``.
+  - **Behavior change** To avoid dimensionality explosion when a feature has two unique values,
+    the default encoder (:class:`OneHotEncoder`) now drops one of the two
+    vectors (see parameter `drop="if_binary"`).
+  - ``fit_transform`` and ``transform`` can now return unencoded features,
+    like the :class:`ColumnTransformer`'s behavior.
+    Previously, a ``RuntimeError`` was raised.
 * **Backward-incompatible change in the SuperVectorizer**: to apply ``remainder``
   to features (with the ``*_transformer`` parameters), the value ``'remainder'``
   must be passed, instead of ``None`` in previous versions.
   ``None`` now indicates that we want to use the default transformer.
 * Support for Python 3.6 and 3.7 has been dropped. Python >= 3.8 is now required.
 * Bumped minimum dependencies:
-  - sklearn>=0.22
+  - sklearn>=0.23
   - scipy>=1.4.0
   - numpy>=1.17.3
   - pandas>=1.2.0